+++ inst/doc/FAQs.R
@@ -1,6 +1,6 @@
 ## ---- echo = FALSE, results = "hide", message = FALSE--------------------
 require("emmeans")
-knitr::opts_chunk$set(collapse = TRUE, fig.width = 4.5)
+knitr::opts_chunk$set(fig.width = 4.5, class.output = "ro")
 options(show.signif.starts = FALSE)
 
 ## ------------------------------------------------------------------------