+++ inst/doc/confidence-intervals.R
@@ -1,6 +1,6 @@
 ## ---- echo = FALSE, results = "hide", message = FALSE--------------------
 require("emmeans")
-knitr::opts_chunk$set(collapse = TRUE, fig.width = 4.5)
+knitr::opts_chunk$set(fig.width = 4.5, class.output = "ro")
 
 ## ------------------------------------------------------------------------
 pigs.lm <- lm(log(conc) ~ source + factor(percent), data = pigs)
@@ -47,11 +47,19 @@ confint(pigs.rg, by = "source")
 #  summary(.Last.value, by = percent)       ### grouped the other way
 
 ## ------------------------------------------------------------------------
+pigsint.lm <- lm(log(conc) ~ source * factor(percent), data = pigs)
+pigsint.rg <- ref_grid(pigsint.lm)
+contrast(pigsint.rg, "consec", simple = "percent")
+
+## ------------------------------------------------------------------------
 pigs.prs.s <- pairs(pigs.emm.s)
 pigs.prs.s
 
 ## ------------------------------------------------------------------------
 test(pigs.prs.s, joint = TRUE)
+
+## ------------------------------------------------------------------------
+joint_tests(pigsint.rg)
 
 ## ------------------------------------------------------------------------
 test(pigs.prs.s, delta = log(1.25), adjust = "none")